Refactored queue-related logic into the QueueManager class, now responsible for attaching queue events for users to the specific queue
