+++ radio.js
@@ -1,6 +1,7 @@
 var express = require('express');
 var socketio = require('socket.io');
 var Queue = require('./server/core/queue.js');
+var QueueManager = require('./server/core/queue-manager.js');
 var Chat = require('./server/core/chat.js');
 
 // App is the express server that serves the static files in public
@@ -16,6 +17,7 @@ var io = socketio.listen(server);
 
 // Start our song queue
 var queue = new Queue(io.sockets);
+var queueManager = new QueueManager(queue);
 var chat = new Chat();
 var usersCount = 0;
 
@@ -26,41 +28,12 @@ io.on('connection', function(socket){
 	io.emit('usersCount', usersCount);
 
 	chat.attachUser(socket);
-
-	// Send current song and current queue to user
-	if(queue.active && ! queue.active.isOver()) {
-		var info = {url_params: queue.active.getVideoUrlParams(), info: queue.active.getInfo()};
-		socket.emit('new_song', info);
-		socket.emit('queue_info', queue.getInfo());
-		socket.emit('related_info', queue.active.relatedVideos);
-	}
-
-	socket.on('new_song', function(data){
-		// Add to queue
-		queue.add(socket, data);
-	});
-
-	socket.on('delete_queued', function(data){
-		// Remove from queue
-		queue.delete(socket, data);
-	});
-
-	socket.on('vote', function(songId) {
-		// Upvote this song
-		queue.addVote(socket, songId);
-	});
+	queueManager.attachUser(socket);
 
 	socket.on('disconnect', function () {
 		usersCount--;
 
 		io.emit('usersCount', usersCount);
-
-		/**
-		 * A user has left which would cause the
-		 * sorting of the queue to potentially
-		 * change, we need to let users know
-		 */
-		queue.triggerOnQueueChanged();
 	});
 
 	if(isAdmin) {
